+++ src/__mocks__/vscode.ts
@@ -1,8 +1,15 @@
-const Uri = jest.fn().mockImplementation(() => {
+const Uri: any = jest.fn().mockImplementation(() => {
     return {
-        fsPath: 'test/path.php'
+        scheme: 'file',
+        authority: '',
+        path: 'test/file/path.php',
+        query: '',
+        fragment: '',
+        fsPath: 'test/file/path.php',
+        toString: jest.fn()
     };
 });
+Uri.joinPath = jest.fn();
 
 const MockTextDocument = jest.fn().mockImplementation(() => {
     return {
@@ -103,11 +110,24 @@ const TextEdit = jest.fn().mockImplementation((range, newContent) => {
     return { range, newContent };
 });
 
+class FileSystemError extends Error {
+    public readonly code: string;
+
+    public constructor(messageOrUri: string | typeof Uri) {
+        super(messageOrUri.toString());
+
+        this.code = '';
+    }
+}
+
 const workspace = {
     onDidChangeConfiguration: jest.fn(),
     onDidChangeWorkspaceFolders: jest.fn(),
     getWorkspaceFolder: jest.fn(),
-    getConfiguration: jest.fn()
+    getConfiguration: jest.fn(),
+    fs: {
+        stat: jest.fn()
+    }
 };
 
 const languages = {
@@ -129,6 +149,7 @@ export {
     CodeAction,
     CodeActionKind,
     TextEdit,
+    FileSystemError,
     workspace,
     languages,
 };